+++ src/Host/FastTodo.SeedData/appsettings.json
@@ -1,9 +1,9 @@
 {
     "ConnectionStrings": {
         "SQLServer": "Server=127.0.0.1, 1433;Database=FastTodo;User ID=SA;Password=MyPass@word;Encrypt=True;TrustServerCertificate=true",
-        "SQLite": "Data Source=Data Source=../../../FastTodo.API/FastTodo.db;" // relative path from /debug/bin folder
+        "SQLite": "Data Source=../FastTodo.API/FastTodo.db;" // relative path from /debug/bin folder
     },
-    "SqlProvider": "SQLServer",
+    "SqlProvider": "SQLite",
     "Logging": {
         "LogLevel": {
             "Default": "Information",
